+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 <!-- PENDING-CHANGES -->
 ### Fixed
 * Sync API now correctly works with deleted entries.
+* The SDK now works correctly when using an access token which does not have access to the `master` environment.
 <!-- /PENDING-CHANGES -->
 
 ## [4.1.0](https://github.com/contentful/contentful.php/tree/4.1.0) (2018-11-28)